Update README.md
This Portfolio is a personal website built using Next.js and Tailwind CSS to showcase my technical expertise, projects, and articles. The site is designed to be visually appealing, responsive, and optimized for performance. It includes a dedicated projects section where I highlight my past work, a blog section for sharing technical insights, and easy-to-access contact links for collaboration opportunities. The website is deployed on Vercel for seamless hosting and continuous deployment.